move dump_as_python to loopy.tools
